requirements.txt
flask==2.3.3
langchain==0.0.350
langchain-google-genai==0.0.5
langchain-community==0.0.11
python-dotenv==1.0.0
tenacity==8.2.3
pymysql==1.1.0
SQLAlchemy==2.0.23